refactor(bots): extract bot_token leaf crate
Move hash/prefix helpers out of bot_id so identity stays identity-only.
bots and macro_authorization share the digest without a crate cycle.
