+++ crates/bot_id/Cargo.toml
@@ -13,7 +13,6 @@ test-utils = []
 cowlike = { path = "../cowlike" }
 nom = { workspace = true }
 serde = { workspace = true }
-sha2 = { workspace = true }
 sqlx = { workspace = true, optional = true }
 thiserror = { workspace = true }
 utoipa = { workspace = true, optional = true }
