+++ crates/bot_token/Cargo.toml
@@ -1,0 +1,9 @@
+[package]
+edition = "2024"
+name = "bot_token"
+publish = false
+version = "0.1.0"
+
+[dependencies]
+sha2 = { workspace = true }
+workspace-hack = { version = "0.1", path = "../workspace-hack" }
