+++ crates/bots/Cargo.toml
@@ -37,6 +37,7 @@ anyhow = { workspace = true }
 async-trait = { workspace = true, optional = true }
 axum = { workspace = true, optional = true }
 bot_id = { path = "../bot_id", features = ["schema"] }
+bot_token = { path = "../bot_token" }
 channels = { path = "../channels", optional = true, default-features = false, features = ["ports"] }
 chrono = { workspace = true }
 entity_access = { path = "../entity_access", optional = true, default-features = false }
