+++ crates/bots/src/domain/ports.rs
@@ -5,7 +5,7 @@ use super::models::{
     BotTokenCandidate, CreateBotRequest, CreateBotTokenRequest, CreateBotTokenResponse,
     CreateChannelScopedBotRequest, CreateChannelScopedBotResponse, PatchBotRequest,
 };
-use bot_id::HashedBotToken;
+use bot_token::HashedBotToken;
 use entity_access::domain::models::{EntityAccessReceipt, MemberParticipantRole};
 use macro_user_id::user_id::MacroUserIdStr;
 use std::future::Future;
